SWDEV-295144 - Change uint64_t to size_t to fix failure on 32-bit opencl.


[ROCm/clr commit: 788ae8595b00879decd4cc5cfa436b4e475317bc]

+++ rocclr/platform/memory.hpp
@@ -651,7 +651,7 @@ class LiquidFlashFile : public RuntimeObject {
 class ArenaMemory: public Buffer {
 public:
   ArenaMemory(Context& context)
-    : Buffer(context, 0, std::numeric_limits<uint64_t>::max(),
+    : Buffer(context, 0, std::numeric_limits<size_t>::max(),
              reinterpret_cast<void*>(kArenaMemoryPtr)) {}
 };
